refactor: 메인 페이지 상품 리스트, 북마크 리스트 렌더링 되지 않던 버그 fix 및 코드 리팩토링
